zdapd ZFS performance improvements
The current ZFS lib opens a dataset including all children, including
snapshots when a dataset is opened. For some operations a dataset or
snapshot where opened multiple times, causing a performance penalty.

What this chane does, is to reduce the number of times a dataset is
opened, and instead pass around the opened dataset to where it's used.