Support clone pools and claim command (zpeedap) (#12)
Extend zdap with a dynamic clone pool per resource, that can be claimed for testing and then automatically discarded. Each pooled clone sets an expiration date as a zfs property.

Each resource that is configured to use a clone pool has a clone pool worker that runs every hour, or when something is added to its buffered (1) channel. The action will:

Expire all clones that are based on a snap older than the latest snap.
Remove expired clones.
Add clones to the pool to reach the configured minimum value of available clones.
Each claim and new snap immediately adds to the clone pool workers queue, and each claim also starts a goroutine that sleeps until its expiration and then adds itself to the queue. This ensures that it runs every time we know some action is needed, but doesn't run all the time and hogs resources/locks. The reason behind the 1 buffer space in the channel is to ensure that any time anyone adds to the queue, a complete action run is guaranteed to happen either immediately or as soon as the current action finishes.

I also removed our replace directive for lib-zfs and upgraded the version to the latest one. This seems to have solved the segfaults we were getting before, when claiming and destroying lots of claims.

It seems that modifying any dataset and opening the entire pool concurrently causes segfaults with both the old and the new lib. (reproduced on current "live" servers by concurrently running two loops attaching/detaching) I've tried solving this by adding a RW lock, where Open() is wrapped with claiming the read lock and all mutations are wrapped in a write lock. All mutations then need to go through zfs.go.

This should probably be extended with better interfaces that wrap the internal zfs stuff, so you really cannot mutate anything from outside zfs.go. Another likely useful extension is to keep a copy of all local zfs state in memory, so we don't need to traverse the datasets unnecessarily. Then once you've found what you're looking for, e.g. a clone, you can open only that dataset and hopefully you then won't need to lock everything.