Fixes issue where zdap-proxy image may be overwritten
Fixes the issue where zdap-proxy image may be overwritten when rebuilding
containers in the dev-environment. That were caused when we add image in the
override, but don't remove the 'build' YAML node. To fix this we add the
'build' YAML node to the override file, and add the '!reset' custom YAML tag
to the node. That will cause docker to remove/clear the 'build' YAML node when
the override docker compose file is merged.